requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
Django==5.1.3
django-cloudinary-storage==0.3.0
django-filter==24.3
djangorestframework==3.15.2
pillow==11.0.0
pytz==2024.2
sqlparse==0.5.2
